requirements.txt
efficientnet_pytorch==0.7.1
geopandas==0.13.0
joblib==1.2.0
lightning==2.0.2
tensorflow==2.10.1
torch==1.13.1
seaborn
natsort

# Anomalib
albumentations==1.3.1
comet_ml==3.33.11
einops==0.7.0
FrEIA==0.2
imgaug==0.4.0
jsonargparse==4.25.0
kornia==0.6.9
matplotlib==3.8.0
numpy==1.26.0
omegaconf==2.3.0
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
openvino==2023.0.0
pandas==2.1.1
Pillow==10.0.1
pytorch_lightning==1.9.5
Requests==2.31.0
scikit_learn==1.0.2
scipy==1.11.3
setuptools==65.5.1
tifffile==2023.9.26
tqdm==4.66.1
urllib3==2.0.6
timm==0.9.7
tensorboard==2.10.1